# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Pocket Gull Node.js Express Gateway",
    "version": "1.0.0",
    "description": "API Gateway for Pocket Gull - handles secrets, patient state sync, and PubMed proxy services."
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local Development Server"
    },
    {
      "url": "/api",
      "description": "Relative URL (Proxied via Frontend)"
    }
  ],
  "paths": {
...
    "/api/pubmed/search": {
      "get": {
        "summary": "PubMed Search Proxy",
        "description": "Queries the NCBI Entrez PubMed database for articles using a search term.",
        "parameters": [
          {
            "name": "term",
            "in": "query",
            "required": true,
            "description": "The search query (e.g., 'hypothyroidism symptoms')",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval of PubMed search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "esearchresult": {
                      "type": "object",
                      "properties": {
                        "count": { "type": "string" },
                        "retmax": { "type": "string" },
                        "retstart": { "type": "string" },
                        "idlist": {
                          "type": "array",
                          "items": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing search term",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Term is required" }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/pubmed/summary": {
      "get": {
        "summary": "PubMed Summary Proxy",
        "description": "Retrieves the detail summaries for a list of PubMed article IDs.",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "description": "Comma-separated PubMed IDs (e.g., '32912445,31023902')",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval of PubMed article summaries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "object",
                        "properties": {
                          "uid": { "type": "string" },
                          "title": { "type": "string" },
                          "pubdate": { "type": "string" },
                          "source": { "type": "string" },
                          "authors": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "name": { "type": "string" },
                                "authtype": { "type": "string" }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "ID is required" }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
...
}